add tests for JWT attacks that should continue to not work
after reading https://www.chosenplaintext.ca/2015/03/31/jwt-algorithm-confusion.html, i'd like
to document and ensure that alg=none and alg=hs256 attacks do not and will continue to not
work.

the alg=hs256 attack involves fetching the public key and using it to generate a JWT as if it
were an HMAC secret. the reason this doesn't work is because my token verification process
assumes RS256 and passes a JWK to the verify!() method, which raises a TypeError from
OpenSSL::HMAC.digest (expected: String).

the alg=none attack involves stripping the signature and pretending like it was never meant to
exist. the reason this doesn't work is because my token verification process calls verify!()
with a JWK, and the library raises an error if a secret is provided when alg=none.